requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
stripe>=7.0.0
